+++ app/components/dashboard/home/index.jsx
@@ -1,0 +1,70 @@
+import { Box, Button, Typography } from "@mui/material";
+import { useTheme } from "@mui/system";
+import Image from "next/image";
+import React from "react";
+import { tokens } from "../../../theme";
+import Ad from "./Ad";
+import AppAd from "./AppAd";
+
+const Home = () => {
+  const theme = useTheme();
+  const colors = tokens(theme.palette.mode);
+
+  return (
+    <Box
+      boxShadow="0 0 5px black"
+      height="fit-content"
+      width="100%"
+      borderRadius="5px"
+      gridColumn="2"
+      gridTemplateColumns="50% 50%"
+      display="grid"
+    >
+      <Ad />
+      <AppAd />
+      <Box
+        boxShadow="0 0 5px black"
+        height="fit-content"
+        // width="fit-content"
+
+        m="30px"
+        borderRadius="5px"
+        gridColumn="span 1"
+      >
+        <Box>Payment box</Box>
+      </Box>
+      <Box
+        boxShadow="0 0 5px black"
+        height="fit-content"
+        // width="fit-content"
+        m="30px"
+        borderRadius="5px"
+        gridColumn="1 / span 2"
+      >
+        <Box>Ad box</Box>
+      </Box>
+      <Box
+        boxShadow="0 0 5px black"
+        height="fit-content"
+        // width="fit-content"
+        m="30px"
+        borderRadius="5px"
+        gridColumn="1 / span 2"
+      >
+        <Box>Ad box</Box>
+      </Box>
+      <Box
+        boxShadow="0 0 5px black"
+        height="fit-content"
+        // width="fit-content"
+        m="30px"
+        borderRadius="5px"
+        gridColumn="1 / span 2"
+      >
+        <Box>Ad box</Box>
+      </Box>
+    </Box>
+  );
+};
+
+export default Home;
